app/requirements.txt
datasets
lightning
peft
torchmetrics
transformers[torch]
flask
gunicorn